+++ tests/MSDFTests.hs
@@ -4,12 +4,16 @@ import Control.Exception (SomeException, try, evaluate)
 import Data.Array (Array, elems, (!))
 import Data.Array.IArray (bounds, rangeSize)
 import qualified Data.Array.Unboxed as UA
-import Data.Bits (testBit, (.&.))
+import Data.Bits (testBit, (.&.), xor, shiftR)
 import Data.List (find)
-import Data.Maybe (fromMaybe)
-import Data.Word (Word16)
+import Data.Maybe (fromMaybe, isNothing)
+import Data.Word (Word8, Word16, Word32, Word64)
+import Foreign.ForeignPtr (withForeignPtr, mallocForeignPtrBytes)
+import Foreign.Marshal.Utils (copyBytes)
+import Foreign.Ptr (Ptr, plusPtr)
+import Foreign.Storable (pokeByteOff)
 import System.Exit (exitFailure)
-import System.IO (hFlush, hPutStrLn, stdout, stderr)
+import System.IO (hFlush, hPutStrLn, stdout, stderr, withBinaryFile, IOMode(WriteMode), hPutBuf)
 
 import MSDF.Generated (generateMSDFWithConfig)
 import MSDF.MSDF
@@ -22,47 +26,88 @@ import MSDF.MSDF
   , GlyphSet(..)
   )
 import qualified MSDF.MSDF as MSDF
-import MSDF.Render (glyphQuad, glyphUV)
-import MSDF.Binary (ByteBuffer, readS16BE, readU16BE, slice)
+import MSDF.Render (glyphQuad, glyphUV, scaleForPixelSize, pixelRangeForAtlas)
+import MSDF.Binary (ByteBuffer(..), readByteBuffer, readS16BE, readU16BE, readU8, readTag, slice)
 import MSDF.Outline (Point(..))
-import MSDF.TTF.GPOS (GPOSMarksRaw(..), MarkToBaseRaw(..), MarkToMarkRaw(..), MarkGlyphRaw(..), BaseGlyphRaw(..))
+import MSDF.TTF.GPOS (AnchorRaw(..), GPOSMarksRaw(..), MarkToBaseRaw(..), MarkToMarkRaw(..), MarkGlyphRaw(..), BaseGlyphRaw(..))
 import MSDF.TTF.Parser
-import MSDF.TTF.Variations (Fvar(..), FvarAxis(..), Gvar(..), applyGvarToContours, hvarDeltas, vvarDeltas, mvarHheaDeltas, mvarVheaDeltas)
+import MSDF.TTF.Variations (Fvar(..), FvarAxis(..), Gvar(..), applyGvarToContours, componentDeltas, hvarDeltas, vvarDeltas, mvarHheaDeltas, mvarVheaDeltas)
 import MSDF.Types
 import Paths_masdiff (getDataFileName)
 
 ttfPath :: IO FilePath
 ttfPath = getDataFileName "assets/Inter/Inter-VariableFont_opsz,wght.ttf"
 
+staticRegularPath :: IO FilePath
+staticRegularPath = getDataFileName "assets/Inter/static/Inter_24pt-Regular.ttf"
+
+staticBoldPath :: IO FilePath
+staticBoldPath = getDataFileName "assets/Inter/static/Inter_24pt-Bold.ttf"
+
+italicVarPath :: IO FilePath
+italicVarPath = getDataFileName "assets/Inter/Inter-Italic-VariableFont_opsz,wght.ttf"
+
 main :: IO ()
 main = do
   fontPath <- ttfPath
   ttf <- requireRight "parseTTF" =<< parseTTF fontPath
+  regPath <- staticRegularPath
+  boldPath <- staticBoldPath
+  italicPath <- italicVarPath
+  ttfRegular <- requireRight "parseTTF regular" =<< parseTTF regPath
+  ttfBold <- requireRight "parseTTF bold" =<< parseTTF boldPath
+  ttfItalic <- requireRight "parseTTF italic" =<< parseTTF italicPath
   let cfgSmall :: MSDFConfig
       cfgSmall = defaultMSDFConfig { MSDF.pixelSize = 16 }
       cfgSubset = cfgSmall { MSDF.glyphSet = GlyphSetCodepoints [65, 66] }
+      cfgUnpacked = cfgSubset { MSDF.packAtlas = False }
+      cfgVarSubset = cfgSmall
+        { MSDF.glyphSet = GlyphSetCodepoints [65, 66, 67, 68]
+        , MSDF.variations = [("wght", 700), ("opsz", 32)]
+        , MSDF.packAtlas = False
+        }
   subsetAtlas <- requireRight "generateMSDFWithConfig" =<< generateMSDFWithConfig cfgSubset fontPath
+  unpackedAtlas <- requireRight "generateMSDFWithConfig" =<< generateMSDFWithConfig cfgUnpacked fontPath
+  varAtlas <- requireRight "generateMSDFWithConfig" =<< generateMSDFWithConfig cfgVarSubset fontPath
   parallelAtlas <- requireRight "generateMSDFWithConfig" =<< generateMSDFWithConfig (cfgSubset { MSDF.parallelism = 16 }) fontPath
   results <- sequence
     [ runTest "parseTTF basic" (testParseTTF ttf)
+    , runTest "static font parse" (testStaticFontParse ttfRegular ttfBold)
+    , runTest "static font differ" (testStaticFontDiffer ttfRegular ttfBold)
     , runTest "cmap includes" (testCmap ttf)
     , runTest "glyph outline A" (testOutlineA ttf)
+    , runTest "repeat flags parsing" (testRepeatFlagsParsing ttf)
     , runTest "composite glyph" (testCompositeGlyph ttf)
     , runTest "render glyph MSDF" (testRenderGlyph ttf cfgSmall)
     , runTest "variable font axes" (testVariableFont ttf)
+    , runTest "variation normalization" (testVariationNormalization ttf)
+    , runTest "italic variable font" (testItalicVariableFont ttfItalic)
+    , runTest "variable glyph batch" (testVariableGlyphBatch ttf)
+    , runTest "variable composite deltas" (testVariableCompositeDeltas ttf)
+    , runTest "variable metrics batch" (testVariableMetricsBatch ttf)
+    , runTest "variable atlas" (testVariableAtlas varAtlas)
     , runTest "gvar default safe" (testGvarDefaultSafe ttf)
     , runTest "gpos marks" (testGposMarks ttf)
     , runTest "subset atlas" (testGenerateSubset subsetAtlas)
     , runTest "atlas packing" (testAtlasPacking subsetAtlas)
+    , runTest "atlas placement bounds" (testAtlasPlacementBounds subsetAtlas)
+    , runTest "atlas power of two" (testAtlasPowerOfTwo subsetAtlas)
     , runTest "render helpers" (testRenderHelpers subsetAtlas)
+    , runTest "render scale helpers" (testRenderScaleHelpers subsetAtlas)
+    , runTest "unpacked atlas" (testUnpackedAtlas unpackedAtlas)
     , runTest "kerning sorted" (testKerningSorted subsetAtlas)
     , runTest "codepoint index sorted" (testCodepointIndexSorted subsetAtlas)
     , runTest "lookup codepoint" (testLookupCodepoint subsetAtlas ttf)
     , runTest "parallelism deterministic" (testParallelismDeterministic ttf subsetAtlas parallelAtlas)
     , runTest "glyph cache" (testGlyphCache ttf cfgSmall)
+    , runTest "bitmap hash" (testBitmapHash ttf cfgSmall)
+    , runTest "bitmap hash variable" (testBitmapHashVariable ttf)
+    , runTest "bitmap hash italic" (testBitmapHashItalic ttfItalic)
+    , runTest "composite variation" (testCompositeVariation ttf)
     , runTest "metrics only" (testMetricsOnly ttf cfgSmall)
     , runTest "vertical metrics" (testVerticalMetrics ttf cfgSmall)
     , runTest "composite point match" (testCompositePointMatch ttf)
+    , runTest "malformed tables" (testMalformedTables fontPath)
     ]
   if and results
     then putStrLn "All tests passed."
@@ -99,6 +144,26 @@ testParseTTF ttf = do
   assert (ttf.maxp.numGlyphs > 0) "numGlyphs should be > 0"
   assert (not (null ttf.cmap.mappings)) "cmap mappings should not be empty"
 
+testStaticFontParse :: TTF -> TTF -> IO ()
+testStaticFontParse regular bold = do
+  assert (isNothing regular.variations) "regular should not have variations"
+  assert (isNothing bold.variations) "bold should not have variations"
+  assert (regular.maxp.numGlyphs > 0) "regular font has no glyphs"
+  assert (bold.maxp.numGlyphs > 0) "bold font has no glyphs"
+
+testStaticFontDiffer :: TTF -> TTF -> IO ()
+testStaticFontDiffer regular bold = do
+  let cps = [65, 66, 67, 68, 69, 77, 97, 109]
+      diffCp =
+        [ cp
+        | cp <- cps
+        , let gR = fromMaybe (-1) (lookupCodepointList cp regular.cmap.mappings)
+        , let gB = fromMaybe (-1) (lookupCodepointList cp bold.cmap.mappings)
+        , gR >= 0 && gB >= 0
+        , contourHash (glyphOutline regular gR) /= contourHash (glyphOutline bold gB)
+        ]
+  assert (not (null diffCp)) "regular and bold outlines identical for sampled glyphs"
+
 
 testCmap :: TTF -> IO ()
 testCmap ttf = do
@@ -118,6 +183,26 @@ testOutlineA ttf = do
   let contours = glyphOutline ttf gA
   assert (not (null contours)) "glyph A has no contours"
   assert (any (not . null) contours) "glyph A contours empty"
+
+testRepeatFlagsParsing :: TTF -> IO ()
+testRepeatFlagsParsing ttf =
+  case findRepeatFlagGlyph ttf of
+    Nothing -> pure ()
+    Just (glyphIndex, bb, endPts, flags, flagsOff, bytesConsumed) -> do
+      let xOff = flagsOff + bytesConsumed
+          (xs, yOff) = readCoordsTest bb xOff flags True
+          (ys, _) = readCoordsTest bb yOff flags False
+          points =
+            [ Point (fromIntegral (xs !! i)) (fromIntegral (ys !! i)) (testBit (flags !! i) 0)
+            | i <- [0 .. min (length flags - 1) (length xs - 1)]
+            ]
+          contoursExpected = splitContoursTest points endPts
+          contoursActual = glyphOutline ttf glyphIndex
+          expectedHash = contourHash contoursExpected
+          actualHash = contourHash contoursActual
+      assert (not (null contoursExpected)) "repeat-flag glyph has no contours"
+      assert (bytesConsumed /= length flags) "repeat-flag glyph did not use compression"
+      assert (expectedHash == actualHash) "repeat-flag glyph contour mismatch"
 
 
 testCompositeGlyph :: TTF -> IO ()
@@ -145,6 +230,10 @@ testGposMarks ttf = do
       GPOSMarksRaw markToBase' markToMark' = marks
   mapM_ (checkMarkToBase numGlyphs) markToBase'
   mapM_ (checkMarkToMark numGlyphs) markToMark'
+  let anchors = concatMap anchorsFromMarkToBase markToBase' ++ concatMap anchorsFromMarkToMark markToMark'
+  if null markToBase' && null markToMark'
+    then pure ()
+    else assert (not (null anchors)) "GPOS mark tables present but no anchors"
 
 checkMarkToBase :: Int -> MarkToBaseRaw -> IO ()
 checkMarkToBase numGlyphs (MarkToBaseRaw classCount marks bases) = do
@@ -184,6 +273,23 @@ assertArrayBounds numGlyphs arr msg =
      then assert (lo > hi) (msg ++ ": expected empty array")
      else assert (lo == 0 && hi == numGlyphs - 1) (msg ++ ": bounds mismatch")
 
+anchorsFromMarkToBase :: MarkToBaseRaw -> [AnchorRaw]
+anchorsFromMarkToBase (MarkToBaseRaw _ marks bases) =
+  markAnchors marks ++ baseAnchors bases
+
+anchorsFromMarkToMark :: MarkToMarkRaw -> [AnchorRaw]
+anchorsFromMarkToMark (MarkToMarkRaw _ marks1 marks2) =
+  markAnchors marks1 ++ baseAnchors marks2
+
+markAnchors :: Array Int (Maybe MarkGlyphRaw) -> [AnchorRaw]
+markAnchors arr = [ anchor | Just (MarkGlyphRaw _ anchor) <- elems arr ]
+
+baseAnchors :: Array Int (Maybe BaseGlyphRaw) -> [AnchorRaw]
+baseAnchors arr =
+  concat
+    [ [ anchor | Just anchor <- elems anchors ]
+    | Just (BaseGlyphRaw anchors) <- elems arr
+    ]
 
 testRenderGlyph :: TTF -> MSDFConfig -> IO ()
 testRenderGlyph ttf cfg = do
@@ -266,6 +372,127 @@ testVariableFont ttf = do
       _glyphDefault = renderGlyphMSDF cfgDefault ttf gA
       _glyphVar = renderGlyphMSDF cfgVar ttf gA
   pure ()
+
+testVariationNormalization :: TTF -> IO ()
+testVariationNormalization ttf =
+  case ttf.variations of
+    Nothing -> pure ()
+    Just vars ->
+      case vars.fvar.axes of
+        [] -> pure ()
+        (axis0:_) -> do
+          let axisSpan = axis0.maxValue - axis0.minValue
+              big = axis0.maxValue + axisSpan * 10
+              VariationLocation coords =
+                normalizeLocation vars.fvar vars.avar [(axis0.tag, big)]
+          assert (length coords == length vars.fvar.axes) "normalized coords length mismatch"
+          assert (all (\c -> c >= -1.01 && c <= 1.01) coords) "normalized coords out of range"
+
+testItalicVariableFont :: TTF -> IO ()
+testItalicVariableFont ttf =
+  case ttf.variations of
+    Nothing -> assert False "italic variable font has no variation tables"
+    Just vars -> do
+      assert (not (null vars.fvar.axes)) "italic fvar has no axes"
+      case vars.gvar of
+        Nothing -> assert False "italic gvar missing"
+        Just gvar -> do
+          let axisMaxes = [ (axis.tag, axis.maxValue) | axis <- vars.fvar.axes ]
+              loc = normalizeLocation vars.fvar vars.avar axisMaxes
+              allCandidates = gvarGlyphsWithData gvar (ttf.maxp.numGlyphs - 1)
+              candidates = take 200 allCandidates ++ take 200 (reverse allCandidates)
+          assert (not (null candidates)) "italic gvar has no glyph data"
+          varies <- anyM (glyphVariesOrPhantom ttf gvar loc) candidates
+          assert varies "italic variable font should alter outline or phantom deltas"
+
+testVariableGlyphBatch :: TTF -> IO ()
+testVariableGlyphBatch ttf =
+  case ttf.variations of
+    Nothing -> pure ()
+    Just vars ->
+      case vars.gvar of
+        Nothing -> assert False "gvar missing"
+        Just gvar -> do
+          let loc = normalizeLocation vars.fvar vars.avar [("wght", 700), ("opsz", 32)]
+              candidates = take 50 (gvarGlyphsWithData gvar (ttf.maxp.numGlyphs - 1))
+          assert (not (null candidates)) "gvar has no glyph data"
+          mapM_ (checkGlyph gvar loc) candidates
+  where
+    checkGlyph gvar loc i = do
+      let baseContours = glyphOutlineAt Nothing ttf i
+          (varContours, (dL, dR, dT, dB)) = applyGvarToContours gvar loc i baseContours
+      assert (length baseContours == length varContours) "gvar changed contour count"
+      assert (and (zipWith sameLen baseContours varContours)) "gvar changed contour point count"
+      assert (all isFinite [dL, dR, dT, dB]) "gvar phantom deltas not finite"
+    sameLen a b = length a == length b
+
+testVariableCompositeDeltas :: TTF -> IO ()
+testVariableCompositeDeltas ttf =
+  case ttf.variations of
+    Nothing -> pure ()
+    Just vars ->
+      case vars.gvar of
+        Nothing -> pure ()
+        Just gvar -> do
+          let loc = normalizeLocation vars.fvar vars.avar [("wght", 700)]
+              maxGlyph = min (ttf.maxp.numGlyphs - 1) (snd (bounds gvar.offsets) - 1)
+              compGlyphs =
+                [ i
+                | i <- [0 .. maxGlyph]
+                , let count = compositeComponentCount ttf i
+                , count > 0
+                , gvarHasData gvar i
+                ]
+              sample = take 20 compGlyphs
+          if null sample
+            then pure ()
+            else mapM_ (checkComposite gvar loc) sample
+  where
+    checkComposite gvar loc i = do
+      let count = compositeComponentCount ttf i
+          (dxs, dys, (dL, dR, dT, dB)) = componentDeltas gvar loc i count
+          (loX, hiX) = bounds dxs
+          (loY, hiY) = bounds dys
+      assert (loX == 0 && hiX == count - 1) "component delta X bounds mismatch"
+      assert (loY == 0 && hiY == count - 1) "component delta Y bounds mismatch"
+      let vals = elems dxs ++ elems dys ++ [dL, dR, dT, dB]
+      assert (all isFinite vals) "component deltas not finite"
+
+testVariableMetricsBatch :: TTF -> IO ()
+testVariableMetricsBatch ttf =
+  case ttf.variations of
+    Nothing -> pure ()
+    Just _ -> do
+      let cfg = defaultMSDFConfig
+            { MSDF.pixelSize = 16
+            , MSDF.variations = [("wght", 700)]
+            }
+          maxGlyph = min (ttf.maxp.numGlyphs - 1) 120
+      mapM_ (checkMetrics cfg ttf) [0 .. maxGlyph]
+  where
+    checkMetrics cfg ttf' i = do
+      let glyph = glyphMetricsOnly cfg ttf' i
+          bb = glyph.bbox
+          vals = [ glyph.advance
+                 , glyph.bearingX
+                 , glyph.bearingY
+                 , bb.xMin
+                 , bb.yMin
+                 , bb.xMax
+                 , bb.yMax
+                 ]
+      assert (all isFinite vals) "metrics contain non-finite values"
+      assert (bb.xMin <= bb.xMax) "bbox xMin > xMax"
+      assert (bb.yMin <= bb.yMax) "bbox yMin > yMax"
+
+testVariableAtlas :: MSDFAtlas -> IO ()
+testVariableAtlas atlas = do
+  assert (isNothing atlas.atlas) "variable atlas should be unpacked"
+  case lookupCodepoint atlas 65 of
+    Nothing -> assert False "variable atlas missing A"
+    Just gA -> do
+      let glyphA = atlas.glyphs ! gA
+      assert (glyphA.bitmap.width > 0) "variable atlas glyph A bitmap empty"
 
 testGvarDefaultSafe :: TTF -> IO ()
 testGvarDefaultSafe ttf = do
@@ -350,6 +577,32 @@ testAtlasPacking atlas = do
             _ -> assert False "glyph A/B placement missing"
         _ -> pure ()
 
+testAtlasPlacementBounds :: MSDFAtlas -> IO ()
+testAtlasPlacementBounds atlas =
+  case atlas.atlas of
+    Nothing -> assert False "atlas packing missing"
+    Just img -> do
+      let glyphs = elems atlas.glyphs
+      mapM_ (checkPlacement img) glyphs
+  where
+    checkPlacement img glyph =
+      case glyph.placement of
+        Nothing -> pure ()
+        Just pl -> do
+          assert (pl.width == glyph.bitmap.width) "placement width mismatch"
+          assert (pl.height == glyph.bitmap.height) "placement height mismatch"
+          assert (pl.x >= 0 && pl.y >= 0) "placement origin out of bounds"
+          assert (pl.x + pl.width <= img.width) "placement width out of bounds"
+          assert (pl.y + pl.height <= img.height) "placement height out of bounds"
+          assert (pl.u0 >= 0 && pl.v0 >= 0 && pl.u1 <= 1 && pl.v1 <= 1) "placement UVs out of range"
+
+testAtlasPowerOfTwo :: MSDFAtlas -> IO ()
+testAtlasPowerOfTwo atlas =
+  case atlas.atlas of
+    Nothing -> assert False "atlas packing missing"
+    Just img -> do
+      assert (isPow2 img.width && isPow2 img.height) "atlas dimensions should be power of two"
+
 testRenderHelpers :: MSDFAtlas -> IO ()
 testRenderHelpers atlas = do
   case lookupCodepoint atlas 65 of
@@ -366,6 +619,26 @@ testRenderHelpers atlas = do
           assert (u0 == 0 && v0 == 0 && u1 == 1 && v1 == 1) "glyphUV default mismatch"
         Just pl -> do
           assert (u0 == pl.u0 && v0 == pl.v0 && u1 == pl.u1 && v1 == pl.v1) "glyphUV placement mismatch"
+
+testRenderScaleHelpers :: MSDFAtlas -> IO ()
+testRenderScaleHelpers atlas = do
+  let px = fromIntegral atlas.pixelSize
+      scale = scaleForPixelSize atlas px
+      range = pixelRangeForAtlas atlas px
+  assertApprox "scaleForPixelSize" 1.0 scale
+  assertApprox "pixelRangeForAtlas" (fromIntegral atlas.range) range
+
+testUnpackedAtlas :: MSDFAtlas -> IO ()
+testUnpackedAtlas atlas = do
+  assert (isNothing atlas.atlas) "unpacked atlas should not include atlas image"
+  let placements = [ pl | glyph <- elems atlas.glyphs, Just pl <- [glyph.placement] ]
+  assert (null placements) "unpacked atlas should not include placements"
+  case lookupCodepoint atlas 65 of
+    Nothing -> pure ()
+    Just gA -> do
+      let glyphA = atlas.glyphs ! gA
+          (u0, v0, u1, v1) = glyphUV glyphA
+      assert (u0 == 0 && v0 == 0 && u1 == 1 && v1 == 1) "glyphUV default mismatch (unpacked)"
 
 
 testKerningSorted :: MSDFAtlas -> IO ()
@@ -430,6 +703,61 @@ testGlyphCache ttf cfg = do
   sigPlain <- glyphSignatureSafe "plain A" glyphPlain
   assertEq "cache compare A" sigPlain sigCached "glyph cache changed output"
 
+testBitmapHash :: TTF -> MSDFConfig -> IO ()
+testBitmapHash ttf cfg = do
+  let mappings = ttf.cmap.mappings
+      gA = fromMaybe (-1) (lookupCodepointList 65 mappings)
+  assert (gA >= 0) "glyph index for A not found"
+  let glyph = renderGlyphMSDF cfg ttf gA
+      h = bitmapHash glyph
+      expected = 0xc6131c6820d910a0 :: Word64
+  assertEq "bitmap hash" expected h ("bitmap hash changed: expected " ++ show expected ++ " got " ++ show h)
+
+testBitmapHashVariable :: TTF -> IO ()
+testBitmapHashVariable ttf = do
+  let mappings = ttf.cmap.mappings
+      gA = fromMaybe (-1) (lookupCodepointList 65 mappings)
+  assert (gA >= 0) "glyph index for A not found"
+  let cfgVar = defaultMSDFConfig
+        { MSDF.pixelSize = 16
+        , MSDF.variations = [("wght", 700), ("opsz", 32)]
+        }
+      glyph = renderGlyphMSDF cfgVar ttf gA
+      h = bitmapHash glyph
+      expected = 0xc6131c6820d910a0 :: Word64
+  assertEq "bitmap hash variable" expected h ("bitmap hash changed: expected " ++ show expected ++ " got " ++ show h)
+
+testBitmapHashItalic :: TTF -> IO ()
+testBitmapHashItalic ttf = do
+  let mappings = ttf.cmap.mappings
+      gA = fromMaybe (-1) (lookupCodepointList 65 mappings)
+  assert (gA >= 0) "glyph index for A not found"
+  let cfgVar = defaultMSDFConfig
+        { MSDF.pixelSize = 16
+        , MSDF.variations = [("wght", 700), ("opsz", 32)]
+        }
+      glyph = renderGlyphMSDF cfgVar ttf gA
+      h = bitmapHash glyph
+      expected = 0xb409680f42ce96b0 :: Word64
+  assertEq "bitmap hash italic" expected h ("bitmap hash changed: expected " ++ show expected ++ " got " ++ show h)
+
+testCompositeVariation :: TTF -> IO ()
+testCompositeVariation ttf =
+  case ttf.variations of
+    Nothing -> pure ()
+    Just vars -> do
+      let numGlyphs = ttf.maxp.numGlyphs
+          compGlyphs = [ i | i <- [0 .. numGlyphs - 1], compositeComponentCount ttf i > 0 ]
+      assert (not (null compGlyphs)) "no composite glyphs found"
+      let locVar = normalizeLocation vars.fvar vars.avar [("wght", 900)]
+          varies i =
+            let baseContours = glyphOutlineAt Nothing ttf i
+                varContours = glyphOutlineAt (Just locVar) ttf i
+            in contourHash baseContours /= contourHash varContours
+      case find varies compGlyphs of
+        Nothing -> assert False "no composite glyph varies under wght=900"
+        Just _ -> pure ()
+
 
 testMetricsOnly :: TTF -> MSDFConfig -> IO ()
 testMetricsOnly ttf cfg = do
@@ -454,6 +782,32 @@ testVerticalMetrics ttf cfg = do
         Just vm -> do
           assert (isFinite vm.advance) "vertical advance must be finite"
           assert (isFinite vm.topSideBearing) "vertical top side bearing must be finite"
+
+testMalformedTables :: FilePath -> IO ()
+testMalformedTables path = do
+  bb <- readByteBuffer path
+  mapM_ (corruptAndExpectFail bb) ["gvar", "GPOS", "HVAR", "MVAR"]
+
+corruptAndExpectFail :: ByteBuffer -> String -> IO ()
+corruptAndExpectFail bb tag =
+  case findTableRecordOffset bb tag of
+    Nothing -> pure ()
+    Just recOff -> do
+      bb' <- copyByteBuffer bb
+      let ByteBuffer { ptr = ptr0, len = len0, off = off0 } = bb'
+          badLen = fromIntegral (len0 + 1024) :: Word32
+      withForeignPtr ptr0 $ \p ->
+        overwriteU32BE (p `plusPtr` off0) (recOff + 12) badLen
+      let tmpPath = "/tmp/masdiff-bad-" ++ tag ++ ".ttf"
+      writeByteBuffer tmpPath bb'
+      result <- parseTTF tmpPath
+      case result of
+        Left _ -> pure ()
+        Right ttf -> do
+          forced <- try (evaluate (forceTable tag ttf)) :: IO (Either SomeException ())
+          case forced of
+            Left _ -> pure ()
+            Right _ -> assert False ("corrupted " ++ tag ++ " table should fail to parse: " ++ tmpPath)
 
 -- Helpers -------------------------------------------------------------------
 
@@ -502,12 +856,97 @@ rectsOverlap a b =
       by1 = b.y + b.height
   in ax0 < bx1 && ax1 > bx0 && ay0 < by1 && ay1 > by0
 
+isPow2 :: Int -> Bool
+isPow2 n = n > 0 && (n .&. (n - 1)) == 0
+
+findTableRecordOffset :: ByteBuffer -> String -> Maybe Int
+findTableRecordOffset bb tag =
+  let numTables = fromIntegral (readU16BE bb 4) :: Int
+      start = 12
+      findRec i
+        | i >= numTables = Nothing
+        | readTag bb (start + i * 16) == tag = Just (start + i * 16)
+        | otherwise = findRec (i + 1)
+  in findRec 0
+
+copyByteBuffer :: ByteBuffer -> IO ByteBuffer
+copyByteBuffer bb = do
+  let ByteBuffer { ptr = srcPtr, len = len0, off = off0 } = bb
+  fptr <- mallocForeignPtrBytes len0
+  withForeignPtr srcPtr $ \src ->
+    withForeignPtr fptr $ \dst ->
+      copyBytes dst (src `plusPtr` off0) len0
+  pure (ByteBuffer fptr len0 0)
+
+writeByteBuffer :: FilePath -> ByteBuffer -> IO ()
+writeByteBuffer path bb =
+  let ByteBuffer { ptr = ptr0, len = len0, off = off0 } = bb
+  in withBinaryFile path WriteMode $ \h ->
+    withForeignPtr ptr0 $ \p ->
+      hPutBuf h (p `plusPtr` off0) len0
+
+overwriteU32BE :: Ptr Word8 -> Int -> Word32 -> IO ()
+overwriteU32BE p off value = do
+  pokeByteOff p off     (fromIntegral (value `shiftR` 24) :: Word8)
+  pokeByteOff p (off+1) (fromIntegral (value `shiftR` 16) :: Word8)
+  pokeByteOff p (off+2) (fromIntegral (value `shiftR` 8) :: Word8)
+  pokeByteOff p (off+3) (fromIntegral value :: Word8)
+
+forceTable :: String -> TTF -> ()
+forceTable tag ttf =
+  case tag of
+    "gvar" -> forceGvar ttf
+    "HVAR" -> forceHvar ttf
+    "MVAR" -> forceMvar ttf
+    "GPOS" -> forceGpos ttf
+    _ -> ()
+
+forceGvar :: TTF -> ()
+forceGvar ttf =
+  case ttf.variations of
+    Nothing -> ()
+    Just vars ->
+      case vars.gvar of
+        Nothing -> ()
+        Just (Gvar _ _ offs _ _ _) ->
+          let (_lo, _hi) = bounds offs
+          in ()
+
+forceHvar :: TTF -> ()
+forceHvar ttf =
+  case ttf.variations of
+    Nothing -> ()
+    Just vars ->
+      case vars.hvar of
+        Nothing -> ()
+        Just hvar -> hvar `seq` ()
+
+forceMvar :: TTF -> ()
+forceMvar ttf =
+  case ttf.variations of
+    Nothing -> ()
+    Just vars ->
+      case vars.mvar of
+        Nothing -> ()
+        Just mvar -> mvar `seq` ()
+
+forceGpos :: TTF -> ()
+forceGpos ttf =
+  let GPOSMarksRaw markToBase' markToMark' = ttf.gposMarks
+  in length ttf.gpos `seq` length markToBase' `seq` length markToMark' `seq` ()
+
 assertEq :: (Eq a, Show a) => String -> a -> a -> String -> IO ()
 assertEq label a b msg = do
   result <- tryAny (evaluate (a == b))
   case result of
     Left e -> error (label ++ ": " ++ show e)
     Right ok -> assert ok msg
+
+assertApprox :: String -> Double -> Double -> IO ()
+assertApprox label expected actual =
+  let eps = 1e-9
+      ok = abs (expected - actual) <= eps
+  in assert ok (label ++ ": expected " ++ show expected ++ " got " ++ show actual)
 
 glyphSignatureSafe :: String -> GlyphMSDF -> IO (Double, Double, Double, BBox, Maybe VerticalMetrics, Int, Int, Int, Maybe GlyphPlacement)
 glyphSignatureSafe label glyph = do
@@ -595,9 +1034,163 @@ glyphVaries ttf loc glyphIndex = do
         Left _ -> pure Nothing
         Right v -> pure (Just v)
 
+glyphVariesOrPhantom :: TTF -> Gvar -> VariationLocation -> Int -> IO Bool
+glyphVariesOrPhantom ttf gvar loc glyphIndex = do
+  let baseContours = glyphOutlineAt Nothing ttf glyphIndex
+  result <- try (evaluate (applyGvarToContours gvar loc glyphIndex baseContours))
+  case result of
+    Left (_ :: SomeException) -> pure False
+    Right (varContours, (dL, dR, dT, dB)) ->
+      let changed = contourHash baseContours /= contourHash varContours
+          phantom = dL /= 0 || dR /= 0 || dT /= 0 || dB /= 0
+      in pure (changed || phantom)
+
+gvarHasData :: Gvar -> Int -> Bool
+gvarHasData gvar i =
+  let offsets = gvar.offsets
+      (lo, hi) = bounds offsets
+  in i >= lo && i + 1 <= hi && offsets ! i < offsets ! (i + 1)
+
+gvarGlyphsWithData :: Gvar -> Int -> [Int]
+gvarGlyphsWithData gvar maxGlyph =
+  let offsets = gvar.offsets
+      (lo, hi) = bounds offsets
+      maxIdx = min maxGlyph (hi - 1)
+  in [ i | i <- [lo .. maxIdx], gvarHasData gvar i ]
+
+findRepeatFlagGlyph :: TTF -> Maybe (Int, ByteBuffer, [Int], [Word8], Int, Int)
+findRepeatFlagGlyph ttf = go 0
+  where
+    numGlyphs = ttf.maxp.numGlyphs
+    Loca { offsets = locaOffsets } = ttf.loca
+    glyf = ttf.glyf
+    go i
+      | i >= numGlyphs = Nothing
+      | otherwise =
+          let start = locaOffsets ! i
+              end = locaOffsets ! (i + 1)
+          in if end <= start
+             then go (i + 1)
+             else
+               let bb = slice glyf start (end - start)
+                   numContours = fromIntegral (readS16BE bb 0) :: Int
+               in if numContours < 0
+                  then go (i + 1)
+                  else
+                    case parseSimpleGlyphFlags bb numContours of
+                      Nothing -> go (i + 1)
+                      Just (endPts, flags, flagsOff, bytesConsumed, usedRepeat) ->
+                        if usedRepeat && bytesConsumed /= length flags
+                        then Just (i, bb, endPts, flags, flagsOff, bytesConsumed)
+                        else go (i + 1)
+
+parseSimpleGlyphFlags :: ByteBuffer -> Int -> Maybe ([Int], [Word8], Int, Int, Bool)
+parseSimpleGlyphFlags bb numContours =
+  let endPtsOff = 10
+      headerBytes = numContours * 2 + 2
+  in if not (withinTest bb endPtsOff headerBytes)
+     then Nothing
+     else
+       let endPts = [ fromIntegral (readU16BE bb (endPtsOff + i * 2))
+                    | i <- [0 .. numContours - 1] ]
+           numPoints = if null endPts then 0 else last endPts + 1
+           instructionLength = fromIntegral (readU16BE bb (endPtsOff + numContours * 2))
+           flagsOff = endPtsOff + numContours * 2 + 2 + instructionLength
+       in if flagsOff > bb.len
+          then Nothing
+          else
+            let (flags, _nextOff, usedRepeat, bytesConsumed) = readFlagsCompressedTest bb flagsOff numPoints
+            in Just (endPts, flags, flagsOff, bytesConsumed, usedRepeat)
+
+readFlagsCompressedTest :: ByteBuffer -> Int -> Int -> ([Word8], Int, Bool, Int)
+readFlagsCompressedTest bb off count =
+  let (flags, next, usedRepeat) = go off count [] False
+  in (reverse flags, next, usedRepeat, next - off)
+  where
+    go idx 0 acc used = (acc, idx, used)
+    go idx n acc used =
+      if not (withinTest bb idx 1)
+      then (acc, idx, used)
+      else
+        let flag = readU8 bb idx
+            hasRepeat = testBit flag 3 && withinTest bb (idx + 1) 1
+            repeatCount = if hasRepeat
+                          then fromIntegral (readU8 bb (idx + 1))
+                          else 0
+            toTake = min n (repeatCount + 1)
+            acc' = replicate toTake flag ++ acc
+            idx' = idx + 1 + if hasRepeat then 1 else 0
+            used' = used || hasRepeat
+        in go idx' (n - toTake) acc' used'
+
+withinTest :: ByteBuffer -> Int -> Int -> Bool
+withinTest bb off size =
+  let ByteBuffer { len = len0 } = bb
+  in off >= 0 && size >= 0 && off + size <= len0
+
+readCoordsTest :: ByteBuffer -> Int -> [Word8] -> Bool -> ([Int], Int)
+readCoordsTest bb off flags isX =
+  go flags off 0 []
+  where
+    go [] idx _ acc = (reverse acc, idx)
+    go (flag:rest) idx lastVal acc =
+      let isShort = testBit flag (if isX then 1 else 2)
+          isSame = testBit flag (if isX then 4 else 5)
+          (delta, idx') =
+            if isShort
+            then
+              if withinTest bb idx 1
+              then
+                let b = fromIntegral (readU8 bb idx)
+                    d = if isSame then b else -b
+                in (d, idx + 1)
+              else (0, idx)
+            else
+              if isSame
+              then (0, idx)
+              else if withinTest bb idx 2
+                   then (fromIntegral (readS16BE bb idx), idx + 2)
+                   else (0, bb.len)
+          val = lastVal + delta
+      in go rest idx' val (val : acc)
+
+splitContoursTest :: [Point] -> [Int] -> [[Point]]
+splitContoursTest points endPts =
+  let indices = map (+1) endPts
+  in go points indices
+  where
+    go _ [] = []
+    go pts (n:ns) =
+      let (a, b) = splitAt n pts
+      in a : go b (map (subtract n) ns)
+
 contourHash :: [[Point]] -> Double
 contourHash contours =
   foldl' (\acc (Point px py _) -> acc + px * 0.7 + py * 0.3) 0 [ p | c <- contours, p <- c ]
+
+bitmapHash :: GlyphMSDF -> Word64
+bitmapHash glyph =
+  let bmp = glyph.bitmap
+      h0 = fnvWord64 fnvOffset (fromIntegral bmp.width)
+      h1 = fnvWord64 h0 (fromIntegral bmp.height)
+  in foldl' fnvByte h1 (UA.elems bmp.pixels)
+
+fnvOffset :: Word64
+fnvOffset = 14695981039346656037
+
+fnvPrime :: Word64
+fnvPrime = 1099511628211
+
+fnvByte :: Word64 -> Word8 -> Word64
+fnvByte h b = (h `xor` fromIntegral b) * fnvPrime
+
+fnvWord64 :: Word64 -> Word64 -> Word64
+fnvWord64 h w =
+  foldl' step h [0 .. 7]
+  where
+    step acc i =
+      let byte = fromIntegral ((w `shiftR` (i * 8)) .&. 0xff)
+      in fnvByte acc byte
 
 isFinite :: Double -> Bool
 isFinite x = not (isNaN x || isInfinite x)